refactor: split analysis/_incremental.py into focused modules
Extract 259-line monolith into four focused modules (all under 100 lines):
- _incremental_context.py: dimension loading and resolution
- _incremental_evidence.py: evidence parsing and fingerprint saving
- _incremental_phases.py: phase execution helpers
- _incremental_orchestrator.py: top-level orchestrator
Keep _incremental.py as a thin re-export hub for backward compatibility.
